requirements.txt
asgiref==3.9.1
Django==5.2.4
django-ckeditor==6.7.3
django-js-asset==3.1.2
numpy==2.3.1
pandas==2.3.1
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
